// generated/src/app/api-models/check-register.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CheckRegister {
  bankAccountCode: string;
  checkNumber: string;
  checkDate: Date;
  seqVendId: number;
  seqVendAddress: number;
  companyCode: string;
  checkAmt: number;
  clearDate: Date;
  checkStatus: string;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  seqCkprtId: number;
  eftTransNumber: string;
  manualCheckFlag: string;
  statChngDate: Date;
  statChngRsnCode: string;
  crossReference: string;
  category: string;
  bulkPayInd: string;
  remittanceId: number;
  bulkAmt: number;
}